requirements.txt
flask 
google-generativeai